backend/requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Werkzeug==2.3.6
zipp==3.16.2
stream-chat==4.9.0
flask-cors==4.0.0
python-dotenv==1.0.0